[Lit] Implement some parser scaffolding. (#3428)
Implement an outline for a simple parser that can handle
function definitions, pass and return statements. This doesn't
handle expressions at all (just skips over their tokens) and
doesn't build any IR yet, but gets some framework in place.

A bad thing about this patch is that it disables the
punctuation-tokens test.  We can either restore a "lex only"
MLIR translation for testing or we can convert it to being
grammatical and tested by the parser.

MODULAR_ORIG_COMMIT_REV_ID: aee0f97b751534ec8aab71f7fe714a57418a0829

+++ KGEN/lib/ParseLit/LitLexer.cpp
@@ -92,7 +92,7 @@ LitToken LitLexer::emitError(const char *loc, const Twine &message) {
 /// Return the indentation level of the specified token.
 /// TODO: Evaluate tracking this inline as part of the lexing loop.  We should
 /// eval how commonly this is queried to figure out the tradeoff.
-Optional<unsigned> LitLexer::getIndentation(const LitToken &tok) const {
+Optional<size_t> LitLexer::getIndentation(const LitToken &tok) const {
   // Count the number of horizontal whitespace characters before the token.
   auto *bufStart = curBuffer.begin();
 
@@ -103,7 +103,7 @@ Optional<unsigned> LitLexer::getIndentation(const LitToken &tok) const {
     return c == '\n' || c == '\r' || c == '\f' || c == '\v';
   };
 
-  unsigned indent = 0;
+  size_t indent = 0;
   const auto *ptr = (const char *)tok.getSpelling().data();
   while (ptr != bufStart && isHorizontalWS(ptr[-1]))
     --ptr, ++indent;
